refactor(obsidian): extract Integrations settings tab from General
Move Local API, Claude Code Skill, and Device Sync into a dedicated
Integrations tab to declutter General settings.
